% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_power_bin_search.R
\name{mlr_power_bin_search}
\alias{mlr_power_bin_search}
\title{Power analysis via binary search}
\usage{
mlr_power_bin_search(
  b1 = 0,
  b2 = 0,
  b3 = 0,
  b4 = 0,
  b5 = 0,
  rx1.x2 = 0,
  rx1.x3 = 0,
  rx1.x4 = 0,
  rx1.x5 = 0,
  rx2.x3 = 0,
  rx2.x4 = 0,
  rx2.x5 = 0,
  rx3.x4 = 0,
  rx3.x5 = 0,
  rx4.x5 = 0,
  desired_power = 0.8,
  alpha = 0.05,
  datasets = 750,
  left = 0,
  right = 1000,
  verbose = TRUE
)
}
\arguments{
\item{b1}{The coefficient you want to be powered to detect.}

\item{b2}{A second beta coefficient.}

\item{b3}{A third beta coefficient.}

\item{b4}{A fourth beta coefficient.}

\item{b5}{The fifth beta coefficient.}

\item{rx1.x2}{The correlation between predictors 1 and 2.}

\item{rx1.x3}{The correlation between predictors 1 and 3.}

\item{rx1.x4}{The correlation between predictors 1 and 4.}

\item{rx1.x5}{The correlation between predictors 1 and 5.}

\item{rx2.x3}{The correlation between predictors 2 and 3.}

\item{rx2.x4}{The correlation between predictors 2 and 4.}

\item{rx2.x5}{The correlation between predictors 2 and 5.}

\item{rx3.x4}{The correlation between predictors 3 and 4.}

\item{rx3.x5}{The correlation between predictors 3 and 5.}

\item{rx4.x5}{The correlation between predictors 4 and 5.}

\item{desired_power}{The desired power for the test of the first beta coefficient.}

\item{alpha}{Type 1 error rate for the significance test of the first beta coefficient.}

\item{datasets}{The number of generated data sets used to calculate power empirically.}

\item{left}{This is the lower bound of the sample size search space.}

\item{right}{This is the upper bound of the sample size search space.}

\item{verbose}{Set equal to TRUE if you want to see a visualization of the search algorithm.}
}
\value{
A list with one element named 'N' which is the sample size needed.
}
\description{
Power analysis via binary search
}
\examples{
mlr_power_bin_search(b1 = 0.3, b2 = 0.1, rx1.x2 = 0.2, desired_power = 0.80, alpha = 0.05)
}
