% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_power_calc.R
\name{mlr_power_calc}
\alias{mlr_power_calc}
\title{Power calculation for coefficient of multiple linear regression model}
\usage{
mlr_power_calc(
  N = 0,
  b1 = 0,
  b2 = 0,
  b3 = 0,
  b4 = 0,
  b5 = 0,
  rx1.x2 = 0,
  rx1.x3 = 0,
  rx1.x4 = 0,
  rx1.x5 = 0,
  rx2.x3 = 0,
  rx2.x4 = 0,
  rx2.x5 = 0,
  rx3.x4 = 0,
  rx3.x5 = 0,
  rx4.x5 = 0,
  alpha = 0.05,
  datasets = 750
)
}
\arguments{
\item{N}{The sample size}

\item{b1}{The coefficient you want to be powered to detect.}

\item{b2}{A second beta coefficient.}

\item{b3}{A third beta coefficient.}

\item{b4}{A fourth beta coefficient.}

\item{b5}{The fifth beta coefficient.}

\item{rx1.x2}{The correlation between predictors 1 and 2.}

\item{rx1.x3}{The correlation between predictors 1 and 3.}

\item{rx1.x4}{The correlation between predictors 1 and 4.}

\item{rx1.x5}{The correlation between predictors 1 and 5.}

\item{rx2.x3}{The correlation between predictors 2 and 3.}

\item{rx2.x4}{The correlation between predictors 2 and 4.}

\item{rx2.x5}{The correlation between predictors 2 and 5.}

\item{rx3.x4}{The correlation between predictors 3 and 4.}

\item{rx3.x5}{The correlation between predictors 3 and 5.}

\item{rx4.x5}{The correlation between predictors 4 and 5.}

\item{alpha}{Type 1 error rate for the significance test of the first beta ceofficient.}

\item{datasets}{The number of generated datasets used to calculate power empirically.}
}
\value{
Returns the power of the test
}
\description{
Power calculation for coefficient of multiple linear regression model
}
\examples{
mlr_power_calc(N = 600, b1 = 0.1, b2 = 0.3, rx1.x2 = 0.25)
}
