% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muvcov.R
\name{muvcov}
\alias{muvcov}
\title{Title}
\usage{
muvcov(
  b1 = 0,
  b2 = 0,
  b3 = 0,
  b4 = 0,
  b5 = 0,
  rx1.x2 = 0,
  rx1.x3 = 0,
  rx1.x4 = 0,
  rx1.x5 = 0,
  rx2.x3 = 0,
  rx2.x4 = 0,
  rx2.x5 = 0,
  rx3.x4 = 0,
  rx3.x5 = 0,
  rx4.x5 = 0
)
}
\arguments{
\item{b1}{The coefficient you want to be powered to detect.}

\item{b2}{A second beta coefficient.}

\item{b3}{A third beta coefficient.}

\item{b4}{A fourth beta coefficient.}

\item{b5}{The fifth beta coefficient.}

\item{rx1.x2}{The correlation between predictors 1 and 2.}

\item{rx1.x3}{The correlation between predictors 1 and 3.}

\item{rx1.x4}{The correlation between predictors 1 and 4.}

\item{rx1.x5}{The correlation between predictors 1 and 5.}

\item{rx2.x3}{The correlation between predictors 2 and 3.}

\item{rx2.x4}{The correlation between predictors 2 and 4.}

\item{rx2.x5}{The correlation between predictors 2 and 5.}

\item{rx3.x4}{The correlation between predictors 3 and 4.}

\item{rx3.x5}{The correlation between predictors 3 and 5.}

\item{rx4.x5}{The correlation between predictors 4 and 5.}
}
\value{
A list containing a vector of means and a matrix of correlations between the variables in the model.
}
\description{
Title
}
\examples{
muvcov(b1 = 0.3, b2 = 0.1, rx1.x2 = 0.2)
}
