% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_power_calc.R
\name{mlr_power_calc}
\alias{mlr_power_calc}
\title{Power of test for first beta coefficient}
\usage{
mlr_power_calc(N = NULL, mus = NULL, S = NULL, alpha = NULL, datasets = NULL)
}
\arguments{
\item{N}{The sample size.}

\item{mus}{Vector of means.}

\item{S}{Correlation matrix of variables.}

\item{alpha}{Type 1 error rate for the significance test of the first beta ceofficient.}

\item{datasets}{The number of generated datasets used to calculate power empirically.}
}
\value{
A numeric vector of length 1 which contains the power of the test.
}
\description{
Power of test for first beta coefficient
}
\examples{
mus = c(0,0,0)
S = matrix(c(1.0,0.2,0.1,
             0.2,1.0,0.3,
             0.1,0.3,1.0), nrow = 3, byrow = TRUE)
mlr_power_calc(N = 150, mus = mus, S = S, alpha = 0.05, datasets = 1000)
}
